refactor: 템플릿 목록 페이지를 리스트 섹션 구조로 재구성
