+++ src/pages/templates/model/index.ts
@@ -1,0 +1,1 @@
+export * from "./useTemplatesPage";
