+++ src/pages/templates/ui/section/index.ts
@@ -1,0 +1,1 @@
+export * from "./TemplateListSection";
